requirements.txt
# # # Python version required: 3.8

# # # Deep Learning Frameworks
# # torch==1.13.1
# # torchvision==0.14.1
# # tensorflow==2.10.0
# # tensorflow-federated==0.33.0

# # # Federated Learning
# # flwr==1.5.0
# # opacus==1.4.0

# # # Privacy & Encryption
# # cryptography>=41.0.2,<42.0.0
# # pycryptodome==3.18.0
# # tenseal==0.3.14

# # # Data Processing
# # numpy==1.24.4
# # pandas==1.5.3
# # scikit-learn==1.2.2
# # imbalanced-learn==0.10.1

# # # Visualization
# # matplotlib==3.6.3
# # seaborn==0.12.2
# # plotly==5.14.0

# # # Web Framework
# # flask==2.2.2
# # flask-cors==3.0.10
# # fastapi==0.95.2
# # uvicorn==0.20.0
# # pydantic==1.10.7

# # # Database
# # sqlalchemy==1.4.46
# # pymongo==4.3.3
# # redis==4.5.4

# # # Communication
# # requests==2.28.2
# # websockets==10.4
# # grpcio==1.46.3
# # protobuf>=3.9.2,<3.20

# # # Configuration
# # pyyaml==6.0
# # python-dotenv==0.21.1
# # hydra-core==1.3.2

# # # Logging & Monitoring
# # loguru==0.6.0
# # tensorboard==2.10.1
# # wandb==0.13.10
# # prometheus-client==0.15.0

# # # Interactive Shell
# # ipython==8.12.3

# # # Security
# # python-jose==3.3.0
# # passlib==1.7.4
# # bcrypt==4.0.0
# # pyotp==2.9.0

# # # Testing
# # pytest==7.4.0
# # pytest-cov==4.1.0
# # pytest-asyncio==0.21.0
# # pytest-mock==3.11.0
# # hypothesis==6.82.0

# # # Code Quality
# # black==23.7.0
# # flake8==6.0.0
# # pylint==2.17.0
# # mypy==1.4.0
# # isort==5.12.0

# # # Documentation
# # sphinx==7.0.0
# # sphinx-rtd-theme==1.3.0
# # mkdocs==1.5.0
# # mkdocs-material==9.1.0

# # # Utilities
# # tqdm==4.65.0
# # joblib==1.3.0
# # click==8.1.0
# # rich==13.5.0
# # python-multipart==0.0.6

# # # Development
# # jupyter==1.0.0
# # notebook==6.5.4
# # jupyterlab==3.6.5

# # # Docker & Deployment
# # gunicorn==21.2.0
# # docker==6.1.0
# # kubernetes==27.2.0

# # # Performance
# # numba==0.57.0
# # cython==3.0.0

# # # Time Series (for fraud patterns)
# # statsmodels==0.14.0
# # prophet==1.1.4

# # # Graph Analysis (for transaction networks)
# # networkx==3.1


# # Python version required: 3.9.x ONLY (Recommended: 3.9.13 or 3.9.16)
# # tensorflow-federated 0.33.0 explicitly requires Python 3.9
# # NOT compatible with Python 3.8, 3.10, 3.11, or 3.12

# # Deep Learning Frameworks
# torch==1.13.1
# torchvision==0.14.1
# torchaudio==0.13.1
# tensorflow==2.10.1
# tensorflow-federated==0.33.0

# # Federated Learning
# flwr==1.5.0
# opacus==1.4.0

# # Privacy & Encryption
# cryptography==41.0.7
# pycryptodome==3.19.0
# tenseal==0.3.14

# # Data Processing (Compatible with tensorflow-federated 0.33.0)
# numpy==1.21.6
# pandas==1.3.5
# scikit-learn==1.0.2
# imbalanced-learn==0.9.1

# # Visualization
# matplotlib==3.5.3
# seaborn==0.11.2
# plotly==5.11.0

# # Web Framework
# flask==2.2.5
# flask-cors==4.0.0
# werkzeug==2.2.3
# fastapi==0.95.2
# uvicorn==0.20.0
# pydantic==1.10.13

# # Database
# sqlalchemy==1.4.51
# pymongo==4.3.3
# redis==4.5.5

# # Communication (Critical: grpcio must be ~1.46.3 for tensorflow-federated 0.33.0)
# requests==2.28.2
# websockets==10.4
# grpcio==1.46.3
# protobuf==3.19.6

# # Configuration
# pyyaml==6.0.1
# python-dotenv==1.0.0
# hydra-core==1.3.2
# omegaconf==2.3.0

# # Logging & Monitoring
# loguru==0.6.0
# tensorboard==2.10.1
# wandb==0.13.10
# prometheus-client==0.15.0

# # Interactive Shell
# ipython==8.5.0

# # Security
# python-jose==3.3.0
# passlib==1.7.4
# bcrypt==4.0.1
# pyotp==2.9.0

# # Testing
# pytest==7.2.2
# pytest-cov==4.0.0
# pytest-asyncio==0.20.3
# pytest-mock==3.10.0
# hypothesis==6.68.2

# # Code Quality
# black==22.12.0
# flake8==6.0.0
# pylint==2.15.10
# mypy==0.991
# isort==5.11.5

# # Documentation
# sphinx==5.3.0
# sphinx-rtd-theme==1.2.0
# mkdocs==1.4.2
# mkdocs-material==9.0.15

# # Utilities
# tqdm==4.64.1
# joblib==1.2.0
# click==8.1.3
# rich==13.3.5
# python-multipart==0.0.6

# # Development
# jupyter==1.0.0
# notebook==6.5.4
# jupyterlab==3.6.5

# # Docker & Deployment
# gunicorn==20.1.0
# docker==6.1.3
# kubernetes==26.1.0

# # Performance
# numba==0.56.4
# cython==0.29.36

# # Time Series (for fraud patterns)
# statsmodels==0.13.5
# prophet==1.1.2

# # Graph Analysis (for transaction networks)
# networkx==2.8.8

# # Additional required dependencies for tensorflow-federated 0.33.0
# absl-py==1.0.0
# attrs==21.2.0
# cachetools==3.1.1
# dm-tree==0.1.8
# farmhashpy==0.4.0
# googleapis-common-protos==1.56.4
# jaxlib==0.4.13
# jax==0.4.13
# portpicker==1.3.9
# scipy==1.7.3
# tensorflow-model-optimization==0.7.5
# tensorflow-privacy==0.8.9

# # Core dependencies
# typing-extensions==4.5.0
# six==1.16.0
# setuptools==65.6.3
# wheel==0.38.4

# Python version required: 3.10.x (Recommended: 3.10.11)
# This version works perfectly on Windows without tensorflow-federated
# Use Flower (flwr) for federated learning instead

# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
# tensorflow==2.12.0

# Federated Learning (Alternative to tensorflow-federated)
flwr>=1.5.0
opacus>=1.4.0

# Privacy & Encryption
cryptography>=41.0.7
pycryptodome>=3.19.0
tenseal>=0.3.14

# Data Processing
numpy>=1.23.5
pandas>=2.0.3
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.5.0
redis>=5.0.0

# Communication
requests>=2.31.0
websockets>=11.0.0
grpcio>=1.56.0
protobuf>=3.20.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0
hydra-core>=1.3.2
omegaconf>=2.3.0

# Logging & Monitoring
loguru>=0.7.0
tensorboard>=2.12.0
wandb>=0.15.0
prometheus-client>=0.17.0

# Interactive Shell
ipython>=8.15.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pyotp>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
hypothesis>=6.87.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
pylint>=2.17.0
mypy>=1.5.0
isort>=5.12.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
click>=8.1.0
rich>=13.5.0
python-multipart>=0.0.6

# Development
jupyter>=1.0.0
notebook>=7.0.0
jupyterlab>=4.0.0

# Docker & Deployment
gunicorn>=21.2.0
docker>=6.1.0
kubernetes>=27.2.0

# Performance
numba>=0.58.0
cython>=3.0.0

# Time Series (for fraud patterns)
statsmodels>=0.14.0
prophet>=1.1.0

# Graph Analysis (for transaction networks)
networkx>=3.1

# Core dependencies
typing-extensions>=4.8.0
six>=1.16.0
setuptools>=68.0.0
wheel>=0.41.0